+++ tmrl/__init__.py
@@ -5,13 +5,10 @@ logging.basicConfig(stream=sys.stdout, level=logging.DEBUG)
 
 # standard library imports
 from dataclasses import dataclass
-from tmrl.networking import Server, RolloutWorker
-from tmrl.training import Trainer
-from tmrl.custom.custom_gym_interfaces import TM2020InterfaceLidar
+# from tmrl.networking import Server, RolloutWorker, Trainer
+# from tmrl.custom.custom_gym_interfaces import TM2020InterfaceLidar
 from tmrl.envs import GenericGymEnv
 from tmrl.config.config_objects import CONFIG_DICT
-from tmrl.tools.record import record_reward_dist
-from tmrl.tools.check_environment import check_env_tm20lidar
 
 
 def get_environment():
